requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
aiosmtplib==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
faker==20.1.0
pydantic[email]==2.5.0
python-multipart==0.0.6
